src/requirements.txt
nebula3-python
Flask==1.1.1
jinja2<3.1.0
itsdangerous==2.0.1
werkzeug==2.0.3
pyahocorasick==1.4.2
click==7.1.2
pyyaml
gunicorn
